requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
pypdf
python-multipart
scikit-learn
networkx
sentence-transformers
transformers
torch
